pyproject.toml
[tool.poetry]
name = "ceiling-calculator-bot"
version = "1.0.0"
description = "Telegram bot для расчёта натяжных потолков"
authors = ["Tigran <your.email@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.13,<3.15"
aiogram = "^3.23.0"
pydantic = "^2.10.0"
pydantic-settings = "^2.7.0"
weasyprint = "^62.3"
jinja2 = "^3.1.5"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.8.4"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.mypy]
python_version = "3.13"
strict = true

